+++ apps/website-new/src/i18n/enUS.ts
@@ -1,0 +1,59 @@
+export const EN_US = {
+  introduction: 'Introduction',
+  quickStart: 'Quick Start',
+
+  // Slogans
+  slogan: 'A Progressive React Framework for modern web development',
+  secondSlogan1: 'Enjoy the development experience.',
+  secondSlogan2: 'Make innovation happen faster.',
+
+  // Features
+  feature1: 'Rust Bundler',
+  featureDesc1: 'Easily switch to Rspack bundler with faster build speed.',
+  feature2: 'Integrated BFF',
+  featureDesc2:
+    'Develop BFF code in the same project, enjoy simple function calls.',
+  feature3: 'Nested Routes',
+  featureDesc3: 'File-as-route, comes with lots performance optimizations.',
+  feature4: 'Multi-Rendering Mode',
+  featureDesc4: 'SSR, SSG, SPR, all out of the box for you.',
+  feature5: 'CSS Solutions',
+  featureDesc5: 'CSS Modules, CSS-in-JS, Tailwind CSS, take your pick.',
+  feature6: 'Easy to Configure',
+  featureDesc6:
+    'Launch with zero configuration, then everything is configurable.',
+
+  // Solutions
+  solutions: 'Solutions',
+  solutionsDesc1: 'A progressive React framework for web development.',
+  solutionsDesc2: 'A powerful solution for npm package development.',
+  solutionsDesc3: 'An Rspack-based build tool for web development.',
+  solutionsDesc4: 'A fast Rspack-based static site generator',
+
+  // Footer
+  guide: 'Guide',
+  topic: 'Topic',
+  help: 'Help',
+  coreConcept: 'Core Concept',
+  basicFeatures: 'Basic Features',
+  advancedFeatures: 'Advanced Features',
+  configuration: 'Configuration',
+  command: 'Command',
+  runtime: 'Runtime',
+  conventions: 'Conventions',
+  microFrontend: 'Micro Frontend',
+  stateManagement: 'State Management',
+  pluginSystem: 'Plugin System',
+  projectGenerator: 'Project Generator',
+  githubDiscussion: 'Github Discussion',
+  changelog: 'Changelog',
+
+  // Showcases
+  doubao: 'Doubao',
+  writingo: 'Writingo',
+  volctrans: 'Volctrans',
+  shidianbaike: 'Shidian Baike',
+  xiaohe: 'Xiaohe',
+  dongchedi: 'Dongchedi',
+  volcengineDeveloper: 'Volcengine Developer',
+} as const;