+++ apps/website-new/theme/index.tsx
@@ -1,0 +1,11 @@
+import Theme from 'rspress/theme';
+import { HomeLayout } from './pages';
+import './index.css';
+
+// eslint-disable-next-line import/export
+export * from 'rspress/theme';
+
+export default {
+  ...Theme,
+  HomeLayout,
+};