+++ packages/runtime/src/index.ts
@@ -77,5 +77,13 @@ export function registerRemotes(
   return FederationInstance.registerRemotes.apply(FederationInstance, args);
 }
 
+export function registerPlugins(
+  ...args: Parameters<FederationHost['registerPlugins']>
+): ReturnType<FederationHost['registerRemotes']> {
+  assert(FederationInstance, 'Please call init first');
+  // eslint-disable-next-line prefer-spread
+  return FederationInstance.registerPlugins.apply(FederationInstance, args);
+}
+
 // Inject for debug
 setGlobalFederationConstructor(FederationHost);